renaming parts of the supervision tree
