+++ packages/sync-service/lib/electric/application.ex
@@ -41,7 +41,7 @@ defmodule Electric.Application do
     # replication connection, starting a connection pool).
     #
     # Once there is a DB connection pool running, Connection.Manager will start the singleton
-    # `Electric.Shapes.Supervisor` which is responsible for starting the shape log collector
+    # `Electric.Replication.Supervisor` which is responsible for starting the shape log collector
     # and individual shape consumer process trees.
     #
     # See the moduledoc in `Electric.Connection.Supervisor` for more info.
