+++ packages/sync-service/lib/electric/connection/supervisor.ex
@@ -1,11 +1,11 @@
 defmodule Electric.Connection.Supervisor do
   @moduledoc """
   The connection supervisor is a rest-for-one supervisor that starts `Connection.Manager`,
-  followed by `Shapes.Supervisor`.
+  followed by `Replication.Supervisor`.
 
   Connection.Manager monitors all of the connection process that it starts and if any one of
   the goes down with a critical error (such as Postgres shutting down), the connection manager
-  itself will shut down. This will cause the shutdown of Shapes.Supervisor, due to the nature
+  itself will shut down. This will cause the shutdown of Replication.Supervisor, due to the nature
   of the rest-for-one supervision strategy, and, since the latter supervisor is started as a
   `temporary` child of the connection supervisor, it won't be restarted until its child spec is
   re-added by a new call to `start_shapes_supervisor/0`.
@@ -14,7 +14,7 @@ defmodule Electric.Connection.Supervisor do
   working DB pool and we only have a DB pool when the Connection.Manager process can see that
   all of its database connections are healthy. Connection.Manager tries to reopen connections
   when they are closed, with an exponential backoff, so it is the first process to know when a
-  connection has been restored and it's also the one that starts Shapes.Supervisor once it
+  connection has been restored and it's also the one that starts Replication.Supervisor once it
   has successfully initialized a database connection pool.
   """
 
@@ -45,7 +45,7 @@ defmodule Electric.Connection.Supervisor do
     child_spec =
       Supervisor.child_spec(
         {
-          Electric.Shapes.Supervisor,
+          Electric.Replication.Supervisor,
           stack_id: stack_id,
           shape_cache: shape_cache_spec,
           log_collector: shape_log_collector_spec
