+++ packages/sync-service/test/electric/replication/shape_log_collector_test.exs
@@ -48,7 +48,7 @@ defmodule Electric.Replication.ShapeLogCollectorTest do
         prepare_tables_fn: fn _, _ -> {:ok, [:ok]} end,
         log_producer: ShapeLogCollector.name(ctx.stack_id),
         stack_id: ctx.stack_id,
-        consumer_supervisor: Electric.Shapes.ConsumerSupervisor.name(ctx.stack_id),
+        consumer_supervisor: Electric.Shapes.DynamicConsumerSupervisor.name(ctx.stack_id),
         registry: registry_name
       ]
